pyproject.toml
[project]
name = "agentic-finance-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "huggingface-hub>=1.3.3",
    "langchain-core==1.2.7",
    "langchain-groq==1.1.1",
    "langchain==1.0.2",
    "langgraph==1.0.1",
    "mlflow>=2.10,<3.0",
    "numpy>=2.2.6",
    "orjson>=3.11.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "sentence-transformers>=0.2.5.1",
    "ta>=0.11.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "transformers>=2.3.0",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
